@@ -45,7 +45,6 @@
     "README.md"
   ],
   "devDependencies": {
-    "@eslint-community/eslint-plugin-eslint-comments": "^4.3.0",
     "@nomicfoundation/hardhat-mocha": "workspace:^3.0.0-next.6",
     "@nomicfoundation/hardhat-node-test-reporter": "workspace:^3.0.0-next.6",
     "@nomicfoundation/hardhat-test-utils": "workspace:^",
@@ -54,21 +53,14 @@
     "@types/deep-eql": "^4.0.2",
     "@types/mocha": ">=10.0.10",
     "@types/node": "^20.14.9",
-    "@typescript-eslint/eslint-plugin": "^7.7.1",
-    "@typescript-eslint/parser": "^7.7.1",
     "c8": "^9.1.0",
-    "eslint": "8.57.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "2.29.1",
-    "eslint-plugin-no-only-tests": "3.1.0",
+    "eslint": "9.25.1",
     "expect-type": "^0.19.0",
     "mocha": "^11.0.0",
     "prettier": "3.2.5",
     "rimraf": "^5.0.5",
     "tsx": "^4.19.3",
-    "typescript": "~5.5.0",
-    "typescript-eslint": "7.7.1"
+    "typescript": "~5.5.0"
   },
   "dependencies": {
     "@nomicfoundation/hardhat-errors": "workspace:^3.0.0-next.6",
