@@ -47,7 +47,6 @@
     "prepack": "pnpm build"
   },
   "devDependencies": {
-    "@eslint-community/eslint-plugin-eslint-comments": "^4.3.0",
     "@istanbuljs/nyc-config-typescript": "1.0.2",
     "@microsoft/api-extractor": "7.49.1",
     "@nomicfoundation/hardhat-test-utils": "workspace:^",
@@ -58,24 +57,16 @@
     "@types/mocha": ">=10.0.10",
     "@types/ndjson": "2.0.1",
     "@types/node": "^20.14.9",
-    "@typescript-eslint/eslint-plugin": "^7.7.1",
-    "@typescript-eslint/parser": "^7.7.1",
     "chai": "^5.1.2",
     "chai-as-promised": "^8.0.0",
     "cross-env": "7.0.3",
-    "eslint": "8.57.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "2.29.1",
-    "eslint-plugin-mocha": "^10.0.0",
-    "eslint-plugin-no-only-tests": "3.1.0",
+    "eslint": "9.25.1",
     "mocha": "^11.0.0",
     "nyc": "15.1.0",
     "prettier": "3.2.5",
     "rimraf": "^5.0.5",
     "ts-node": "10.9.2",
-    "typescript": "~5.5.0",
-    "typescript-eslint": "7.7.1"
+    "typescript": "~5.5.0"
   },
   "dependencies": {
     "@ethersproject/address": "5.6.1",
