@@ -18,7 +18,6 @@
     "prepack": "pnpm build"
   },
   "devDependencies": {
-    "@eslint-community/eslint-plugin-eslint-comments": "^4.3.0",
     "@fontsource/roboto": "^5.0.8",
     "@nomicfoundation/ignition-core": "workspace:^3.0.0-next.6",
     "@types/chai": "^4.2.0",
@@ -27,18 +26,10 @@
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "@types/styled-components": "5.1.26",
-    "@typescript-eslint/parser": "^7.7.1",
     "@vitejs/plugin-react": "^4.0.0",
     "chai": "^5.1.2",
     "chai-as-promised": "^8.0.0",
-    "eslint": "8.57.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "2.29.1",
-    "eslint-plugin-no-only-tests": "3.1.0",
-    "eslint-plugin-mocha": "^10.0.0",
-    "eslint-plugin-react-hooks": "^5.1.0",
-    "eslint-plugin-react-refresh": "^0.4.18",
+    "eslint": "9.25.1",
     "mermaid": "10.9.3",
     "mocha": "^11.0.0",
     "prettier": "3.2.5",
