@@ -44,22 +44,15 @@
     "@nomicfoundation/hardhat-viem": "workspace:^3.0.0-next.6",
     "@nomicfoundation/hardhat-test-utils": "workspace:^",
     "@types/node": "^20.14.9",
-    "@typescript-eslint/eslint-plugin": "^7.7.1",
-    "@typescript-eslint/parser": "^7.7.1",
     "c8": "^9.1.0",
     "cross-env": "7.0.3",
-    "eslint": "8.57.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "2.29.1",
-    "eslint-plugin-no-only-tests": "3.1.0",
+    "eslint": "9.25.1",
     "nyc": "15.1.0",
     "prettier": "3.2.5",
     "rimraf": "^5.0.5",
     "ts-node": "10.9.2",
     "tsx": "^4.19.3",
     "typescript": "~5.5.0",
-    "typescript-eslint": "7.7.1",
     "viem": "^2.21.42"
   },
   "dependencies": {
