@@ -59,18 +59,10 @@
     "@types/node": "^20.14.9",
     "@types/prompts": "^2.4.2",
     "@types/sinon": "^10.0.13",
-    "@typescript-eslint/eslint-plugin": "^7.7.1",
-    "@typescript-eslint/parser": "^7.7.1",
     "chai": "^5.1.2",
     "chai-as-promised": "^8.0.0",
     "cross-env": "7.0.3",
-    "eslint": "8.57.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "2.29.1",
-    "eslint-plugin-no-only-tests": "3.1.0",
-    "eslint-plugin-mocha": "^10.0.0",
-    "eslint-plugin-prettier": "4.0.0",
+    "eslint": "9.25.1",
     "mocha": "^11.0.0",
     "nyc": "15.1.0",
     "prettier": "3.2.5",
@@ -78,7 +70,6 @@
     "sinon": "^14.0.0",
     "ts-node": "10.9.2",
     "typescript": "~5.5.0",
-    "typescript-eslint": "7.7.1",
     "viem": "^2.21.42"
   },
   "dependencies": {
