@@ -87,19 +87,19 @@ describe("deploy converts ignition artifact contract to viem instance", () => {
 
   it("should enforce the type is constrained to the contracts functions", async function () {
     await assertRejects(
-      // @ts-expect-error
+      // @ts-expect-error -- Expect an error
       result.externallyLoadedContract.write.nonexistantWrite(),
     );
     await assertRejects(
-      // @ts-expect-error
+      // @ts-expect-error -- Expect an error
       result.externallyLoadedContract.read.nonexistantRead(),
     );
     await assertRejects(
-      // @ts-expect-error
+      // @ts-expect-error -- Expect an error
       result.externallyLoadedContract.estimateGas.nonexistantEstimate(),
     );
     await assertRejects(
-      // @ts-expect-error
+      // @ts-expect-error -- Expect an error
       result.externallyLoadedContract.simulate.nonexistantEstimate(),
     );
   });
